+++ DomainManagement/functions/objectcategories/Resolve-DMObjectCategory.ps1
@@ -22,13 +22,13 @@
 		Resolves the object categories that apply to $adobject
 	#>
 	[CmdletBinding()]
-	Param (
+	param (
 		[Parameter(Mandatory = $true)]
 		$ADObject,
-
+		
 		[PSFComputer]
 		$Server,
-
+		
 		[PSCredential]
 		$Credential
 	)
@@ -40,15 +40,31 @@
 	}
 	process
 	{
-		if ($script:objectCategories.Values.ObjectClass -notcontains $ADObject.ObjectClass) {
+		if ($script:objectCategories.Values.ObjectClass -notcontains $ADObject.ObjectClass)
+		{
 			return
 		}
-
+		
 		$filteredObjectCategories = $script:objectCategories.Values | Where-Object ObjectClass -eq $ADobject.ObjectClass
 		$propertyNames = $filteredObjectCategories.Property | Select-Object -Unique
 		$adObjectReloaded = Get-Adobject @parameters -Identity $ADObject.DistinguishedName -Properties $propertyNames
-		foreach ($filteredObjectCategory in $filteredObjectCategories) {
-			if ($filteredObjectCategory.Testscript.Invoke($adObjectReloaded)) {
+		:main foreach ($filteredObjectCategory in $filteredObjectCategories)
+		{
+			#region Consider Searchbase
+			$resolvedBase = Resolve-String -Text $filteredObjectCategory.SearchBase @parameters
+			switch ($filteredObjectCategory.SearchScope)
+			{
+				'Base' { if ($adObjectReloaded.DistinguishedName -ne $resolvedBase) { continue main } }
+				'OneLevel'
+				{
+					if ($adObjectReloaded.DistinguishedName -notlike "*,$resolvedBase") { continue main }
+					if (($adObjectReloaded.DistinguishedName -split ",").Count -ne (($resolvedBase -split ",").Count + 1)) { continue main }
+				}
+				'Subtree' { if ($adObjectReloaded.DistinguishedName -notlike "*,$resolvedBase") { continue main } }
+			}
+			#endregion Consider Searchbase
+			if ($filteredObjectCategory.Testscript.Invoke($adObjectReloaded))
+			{
 				$filteredObjectCategory
 			}
 		}